+++ packages/desktop-shell/forge.config.js
@@ -1,17 +1,28 @@
+const { customProtocolName } = require('./src/config');
+
 module.exports = {
   prune: false,
   packagerConfig: {
     ignore: [
       'public[/\\\\]*.*',
       'src[/\\\\]*.*',
-      'node_modules[/\\\\]*.*',
+      '^/node_modules[/\\\\]*.*',
       '.gitignore',
       'forge.config.js',
       'project.json',
       'README.md',
+      'CHANGELOG.md',
       'rollup.config.js',
-      'tsconfig.json'
+      'tsconfig.json',
+      'yarn.lock',
+      '.vscode',
+      '.yarn',
     ],
+    asar: true,
+    protocols: [{
+      name: customProtocolName,
+      schemes: [customProtocolName]
+    }]
   },
   makers: []
 }